% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_editor_server.R
\name{model_editor_server}
\alias{model_editor_server}
\title{Shiny UI for Resilient Community Model Editor Tab}
\usage{
model_editor_server(id)
}
\arguments{
\item{id}{Application identifier for namespace.}
}
\description{
Shiny UI for Resilient Community Model Editor Tab
}
\seealso{
Other Controller Modules: 
\code{\link{model_editor_ui}()}
}
\concept{Controller Modules}
