% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_editor_ui.R
\name{model_editor_ui}
\alias{model_editor_ui}
\title{Shiny UI for Resilient Community Model Editor Tab}
\usage{
model_editor_ui(id)
}
\arguments{
\item{id}{Application identifier.}
}
\description{
Smaller left panel with:
\enumerate{
\item selector for team (TODO)
}
\itemize{
\item add/delete team control
}
\enumerate{
\item model file chooser (upload xmile file)
}
\itemize{
\item enabled once a team is selected
}
\enumerate{
\item high level control sliders
}
\itemize{
\item subset of editable parameters?
}
\enumerate{
\item download button to export updated xmile file
}
}
\details{
Table view on right with all editable parameters for each model
}
\seealso{
Other Controller Modules: 
\code{\link{model_editor_server}()}
}
\concept{Controller Modules}
