% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsfeed_server.R
\name{newsfeed_server}
\alias{newsfeed_server}
\title{Shiny server method for Resilient Community Newsfeed page}
\usage{
newsfeed_server(id, reactive_feed = NULL)
}
\arguments{
\item{id}{Application identifier for namespace.}

\item{reactive_feed}{Override persistent storage with a reactive object}
}
\description{
Intended to be used in newsfeed_app, player_app, and control_app.
TODO: replace data list with a table, modify viewer accordingly
}
\seealso{
Other Newsfeed Modules: 
\code{\link{newsfeed_ui}()}
}
\concept{Newsfeed Modules}
