% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsfeed_ui.R
\name{newsfeed_ui}
\alias{newsfeed_ui}
\title{Shiny UI for Resilient Community Newsfeed page}
\usage{
newsfeed_ui(id)
}
\arguments{
\item{id}{Application identifier for namespace.}
}
\description{
Intended to be used in newsfeed_app, player_app, and control_app.
}
\seealso{
Other Newsfeed Modules: 
\code{\link{newsfeed_server}()}
}
\concept{Newsfeed Modules}
