% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_data.R
\name{persistent_data_path}
\alias{persistent_data_path}
\title{Path to storage file for persistent data.}
\usage{
persistent_data_path(fname)
}
\arguments{
\item{fname}{Name of the file.}
}
\value{
Path to the file.  Still may not exist, but directory should.
}
\description{
Persistent data file goes in data directory specified in options.
Create if needed.
}
\seealso{
Other Shared Data Methods: 
\code{\link{persistent_data_watcher}()},
\code{\link{persistent_data_write}()},
\code{\link{persistent_server}()},
\code{\link{persistent_ui}()}
}
\concept{Shared Data Methods}
