% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_data.R
\name{persistent_data_watcher}
\alias{persistent_data_watcher}
\title{Reactive RData file reader with a default}
\usage{
persistent_data_watcher(fname, reactive_override, ...)
}
\arguments{
\item{fname}{Name of file to watch}

\item{reactive_override}{If not NULL, override file watch with this object}

\item{...}{Other args written as default if file does not exist.}
}
\value{
Reactive file reader for the file, or the reactive override.
}
\description{
This wrapper for reactiveFileReader provides default data if no file exists.
Variable args (defaults) written as list to the file on function call if the
file does not exist.
}
\details{
We also automatically place the file in the packages "output" directory.

NOTE: we do NOT compress the file (expecting small files and speed is good)
}
\seealso{
Other Shared Data Methods: 
\code{\link{persistent_data_path}()},
\code{\link{persistent_data_write}()},
\code{\link{persistent_server}()},
\code{\link{persistent_ui}()}
}
\concept{Shared Data Methods}
