% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_data.R
\name{persistent_data_write}
\alias{persistent_data_write}
\title{Write an object to persistent data file.}
\usage{
persistent_data_write(fname, data_obj)
}
\arguments{
\item{fname}{Name of the file.}

\item{data_obj}{Data to write.  No enforcement of correctness here.}
}
\description{
WARNING!  Completely overwrites the file.
}
\seealso{
Other Shared Data Methods: 
\code{\link{persistent_data_path}()},
\code{\link{persistent_data_watcher}()},
\code{\link{persistent_server}()},
\code{\link{persistent_ui}()}
}
\concept{Shared Data Methods}
