% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_server.R
\name{persistent_server}
\alias{persistent_server}
\title{Server for one of our persistent data modules.}
\usage{
persistent_server(id, file_name, reactive_data = NULL, ...)
}
\arguments{
\item{id}{Application id for running app.}

\item{file_name}{Name of file for persistent storage.}

\item{reactive_data}{Optional reactive object for Editor mode.}

\item{...}{Other args passed through as defaults to data_file_watcher.}
}
\description{
This is a component for use in either the status or newsfeed module.
}
\details{
Set up Consumer or Editor mode based on whether a reactive data object is
passed in (Editor handles data in memory, Consumers watch a file).

Observe events on the Publish button (will be enabled in Editor mode only),
and write current data out to file in that case.
}
\seealso{
Other Shared Data Methods: 
\code{\link{persistent_data_path}()},
\code{\link{persistent_data_watcher}()},
\code{\link{persistent_data_write}()},
\code{\link{persistent_ui}()}
}
\concept{Shared Data Methods}
