% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_ui.R
\name{persistent_ui}
\alias{persistent_ui}
\title{UI for one of our persistent data modules.}
\usage{
persistent_ui(id, parent_ns)
}
\arguments{
\item{id}{Application id for running app.}

\item{parent_ns}{Namespace function of the parent module (must be composed).}
}
\description{
This is a component for use in either the status or newsfeed module.
}
\details{
This consists of a "Publish" action button which is hidden unless the module
is in editor mode.

Note on namespaces: because this module is contained within one of the
newsfeed or status modules, which in turn is contained within the app, the
moduleServer method composes those namespaces.  Thus ui elements need to
use a composed namespace in order to match.  For example, in the controller
app status panel, output$persistentMode set in persistent_server will be
visibile to the js code as \code{output['controller-status-persistentMode']}.
}
\seealso{
Other Shared Data Methods: 
\code{\link{persistent_data_path}()},
\code{\link{persistent_data_watcher}()},
\code{\link{persistent_data_write}()},
\code{\link{persistent_server}()}
}
\concept{Shared Data Methods}
