% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_controller_app.R
\name{run_controller_app}
\alias{run_controller_app}
\title{Run the controller shiny app}
\usage{
run_controller_app()
}
\description{
Run the controller shiny app
}
\seealso{
Other Shiny App Runners: 
\code{\link{run_newsfeed_app}()},
\code{\link{run_player_app}()},
\code{\link{run_status_app}()}
}
\concept{Shiny App Runners}
