% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_newsfeed_app.R
\name{run_newsfeed_app}
\alias{run_newsfeed_app}
\title{Run the newsfeed shiny app}
\usage{
run_newsfeed_app()
}
\description{
Run the newsfeed shiny app
}
\seealso{
Other Shiny App Runners: 
\code{\link{run_controller_app}()},
\code{\link{run_player_app}()},
\code{\link{run_status_app}()}
}
\concept{Shiny App Runners}
