% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_app.R
\name{run_one_app}
\alias{run_one_app}
\title{Run a shiny app with options}
\usage{
run_one_app(app_name, data_path = NA)
}
\arguments{
\item{app_name}{Name of application to run (directory under apps)}

\item{data_path}{Path to use for shared data}
}
\description{
This sets up options, finds application dir, and calls shiny::runApp
}
