% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_player_app.R
\name{run_player_app}
\alias{run_player_app}
\title{Run the player shiny app}
\usage{
run_player_app()
}
\description{
Run the player shiny app
}
\seealso{
Other Shiny App Runners: 
\code{\link{run_controller_app}()},
\code{\link{run_newsfeed_app}()},
\code{\link{run_status_app}()}
}
\concept{Shiny App Runners}
