% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_status_app.R
\name{run_status_app}
\alias{run_status_app}
\title{Run the status shiny app}
\usage{
run_status_app()
}
\description{
Run the status shiny app
}
\seealso{
Other Shiny App Runners: 
\code{\link{run_controller_app}()},
\code{\link{run_newsfeed_app}()},
\code{\link{run_player_app}()}
}
\concept{Shiny App Runners}
