% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_server.R
\name{status_server}
\alias{status_server}
\title{Shiny server method for Resilient Community Status page}
\usage{
status_server(id, reactive_status = NULL)
}
\arguments{
\item{id}{Application identifier for namespace.}

\item{reactive_status}{Override persistent storage with a reactive object}
}
\description{
Intended to be used in status_app, player_app, and control_app.
}
\seealso{
Other Status Modules: 
\code{\link{status_ui}()}
}
\concept{Status Modules}
