% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_ui.R
\name{status_ui}
\alias{status_ui}
\title{Shiny UI for Resilient Community Overview page}
\usage{
status_ui(id)
}
\arguments{
\item{id}{Application identifier for namespace.}
}
\description{
Intended to be used in status_app, player_app, and control_app.
}
\seealso{
Other Status Modules: 
\code{\link{status_server}()}
}
\concept{Status Modules}
